pyproject.toml
[tool.poetry]
name = "BazaPoolServices"
version = "0.1.0"
description = "Baza pool services landing, voting backend and payment system"
authors = ["Prasanta Kakati <prasantakakati@baza.foundation>"]
license = "Not open source"

[tool.poetry.dependencies]
python = "^3.7"
django = "^3.0.2"
uwsgi = "^2.0.18"
djangorestframework = "^3.11.1"
django-rest-knox = "^4.1.0"
django-versatileimagefield = "^2.0"
django-cors-headers = "^3.5.0"

[tool.poetry.dev-dependencies]
pep8 = "^1.7.1"
autopep8 = "^1.5"
ipython = "^7.11.1"
flake8 = "^3.7.9"
python-language-server = {extras = ["all"], version = "^0.31.7"}

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
